+++ src/main/java/com/lokoko/domain/user/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.lokoko.domain.user.repository;
+package com.lokoko.domain.user.domain.repository;
 
-import com.lokoko.domain.user.entity.User;
+import com.lokoko.domain.user.domain.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
