+++ src/main/java/com/lokoko/global/auth/service/AuthService.java
@@ -11,8 +11,8 @@ import static com.lokoko.global.utils.LineConstants.PARAM_UI_LOCALES;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.lokoko.domain.user.entity.User;
-import com.lokoko.domain.user.repository.UserRepository;
+import com.lokoko.domain.user.domain.entity.User;
+import com.lokoko.domain.user.domain.repository.UserRepository;
 import com.lokoko.global.auth.entity.enums.OauthLoginStatus;
 import com.lokoko.global.auth.exception.ErrorMessage;
 import com.lokoko.global.auth.exception.OauthException;
